Removed obsolet components, renamed components, added file header
