+++ scanhub-ui/src/LoginContext.tsx
@@ -1,3 +1,10 @@
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * LoginContext.tsx contains the user context.
+ */
+
 import React from 'react';
 import { createContext } from 'react';
