+++ scanhub-ui/src/LoginContextProvider.tsx
@@ -1,3 +1,10 @@
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * LoginContextProvider.tsx contains the user context provider.
+ */
+
 import { useState, PropsWithChildren } from 'react';
 import React from 'react';
