+++ scanhub-ui/src/Routes.tsx
@@ -1,6 +1,10 @@
-// Copyright (C) 2023, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
-// Routes.tsx is responsible for defining the routes of the react app.
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * Routes.tsx is responsible for defining the routes of the react app.
+ */
+
 import React from 'react';
 import { useContext } from 'react';
 import { Route, Routes, Navigate, useNavigate } from 'react-router-dom'
