--- scanhub-ui/src/client/urls.tsx
@@ -1,12 +1,0 @@
-// Copyright (C) 2023, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
-
-// Definition of global variables
-
-const baseUrls = {
-  patientService: 'http://localhost:8100',
-  examService: 'http://localhost:8080',
-  userloginService: 'http://localhost:8080'
-}
-
-export default baseUrls
