+++ scanhub-ui/src/utils/Healthcheck.tsx
@@ -1,6 +1,10 @@
-// Copyright (C) 2023, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
-// Custom hook for health check
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * Healthcheck.tsx contains a custom hook for health check
+ */
+
 import { useEffect, useState } from 'react'
 
 function useHealthCheck(url: string, interval: number = 5000) {
