+++ scanhub-ui/src/utils/Urls.tsx
@@ -1,0 +1,14 @@
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * Urls.tsx contains the different urls of the backend services.
+ */
+
+const baseUrls = {
+  patientService: 'http://localhost:8100',
+  examService: 'http://localhost:8080',
+  userloginService: 'http://localhost:8080'
+}
+
+export default baseUrls
