+++ scanhub-ui/src/utils/Versions.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * Versions.tsx contains the scanhub-ui version from package.json.
+ */
+
+export { version } from '../../package.json'
