--- scanhub-ui/src/utils/size_vars.tsx
@@ -1,15 +1,0 @@
-// Copyright (C) 2023, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
-
-// Definition of global variables
-
-export const navigation = {
-  height: '64px',
-}
-
-export const patientView = {
-  drawerWidth: '400px',
-  procedureListWidth: '300px',
-  iconButtonSize: '30px',
-  iconButtonVariant: 'soft',
-}
