--- scanhub-ui/src/utils/version.ts
@@ -1,1 +1,0 @@
-export { version } from '../../package.json'
