+++ scanhub-ui/src/views/TemplatesView.tsx
@@ -1,6 +1,11 @@
-// Copyright (C) 2023, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
-// PatientTable.tsx is responsible for rendering the patient table view.
+/**
+ * Copyright (C) 2024, BRAIN-LINK UG (haftungsbeschränkt). All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-ScanHub-Commercial
+ * 
+ * TemplatesView.tsx is responsible for rendering all existing template items
+ * and allows to add new templates or edit existing templates.
+ */
+
 import * as React from 'react'
 import Grid from '@mui/joy/Grid'
 
@@ -8,7 +13,7 @@ import ExamTemplateList from '../components/ExamTemplateList'
 import WorkflowTemplateList from '../components/WorkflowTemplateList'
 import TaskTemplateList from '../components/TaskTemplateList'
 
-import { navigation } from '../utils/size_vars'
+import { navigation } from '../utils/SizeVars'
 
 export default function Templates() {
